+++ BE/src/auth/auth.module.ts
@@ -5,11 +5,14 @@ import { AuthService } from "./auth.service";
 import { JwtModule } from "@nestjs/jwt";
 import { PassportModule } from "@nestjs/passport";
 import { JwtStrategy } from "./jwt.strategy";
-import { UsersModule } from "src/users/users.module";
 import { PrivateDiaryGuard } from "./guard/auth.diary-guard";
+import { TypeOrmModule } from "@nestjs/typeorm";
+import { User } from "./users.entity";
+import { UsersRepository } from "./users.repository";
 
 @Module({
   imports: [
+    TypeOrmModule.forFeature([User]),
     PassportModule.register({ defaultStrategy: "jwt" }),
     JwtModule.register({
       secret: process.env.JWT_SECRET,
@@ -17,10 +20,9 @@ import { PrivateDiaryGuard } from "./guard/auth.diary-guard";
         expiresIn: process.env.JWT_ACCESS_TOKEN_TIME,
       },
     }),
-    UsersModule,
   ],
   controllers: [AuthController],
-  providers: [AuthService, JwtStrategy, PrivateDiaryGuard],
-  exports: [PassportModule],
+  providers: [AuthService, JwtStrategy, UsersRepository, PrivateDiaryGuard],
+  exports: [PassportModule, UsersRepository],
 })
 export class AuthModule {}
