+++ BE/src/shapes/shapes.module.ts
@@ -1,6 +1,5 @@
 import { Module } from "@nestjs/common";
 import { TypeOrmModule } from "@nestjs/typeorm";
-import { UsersModule } from "src/users/users.module";
 import { Shape } from "./shapes.entity";
 import { ShapesController } from "./shapes.controller";
 import { ShapesService } from "./shapes.service";
@@ -8,7 +7,7 @@ import { ShapesRepository } from "./shapes.repository";
 import { AuthModule } from "src/auth/auth.module";
 
 @Module({
-  imports: [TypeOrmModule.forFeature([Shape]), UsersModule, AuthModule],
+  imports: [TypeOrmModule.forFeature([Shape]), AuthModule],
   controllers: [ShapesController],
   providers: [ShapesService, ShapesRepository],
   exports: [ShapesRepository],
