--- BE/src/users/users.controller.ts
@@ -1,16 +1,0 @@
-import { Body, Controller, Post, ValidationPipe } from "@nestjs/common";
-import { UsersService } from "./users.service";
-import { CreateUserDto } from "./users.dto";
-
-@Controller("users")
-export class UsersController {
-  constructor(private usersService: UsersService) {}
-
-  @Post()
-  async registerUser(
-    @Body(ValidationPipe) createUserDto: CreateUserDto,
-  ): Promise<void> {
-    await this.usersService.registerUser(createUserDto);
-    return;
-  }
-}
