--- BE/src/users/users.module.ts
@@ -1,14 +1,0 @@
-import { Module } from "@nestjs/common";
-import { TypeOrmModule } from "@nestjs/typeorm";
-import { UsersController } from "./users.controller";
-import { UsersService } from "./users.service";
-import { User } from "./users.entity";
-import { UsersRepository } from "./users.repository";
-
-@Module({
-  imports: [TypeOrmModule.forFeature([User])],
-  controllers: [UsersController],
-  providers: [UsersService, UsersRepository],
-  exports: [UsersRepository],
-})
-export class UsersModule {}
